fix(example): change tryitout to home
closes #156
