--- example/lib/consent/model.dart
@@ -1,6 +1,0 @@
-import 'package:tiki_sdk_flutter/main.dart';
-
-class ConsentExampleModel {
-  ConsentModel? consent;
-  bool isConsentGiven = false;
-}
