--- example/lib/ownership/controller.dart
@@ -1,16 +1,0 @@
-import 'package:example/ownership/service.dart';
-import 'package:example/wallet/service.dart';
-import 'package:flutter/cupertino.dart';
-import 'package:provider/provider.dart';
-import 'package:tiki_sdk_flutter/main.dart';
-
-import '../destination/model.dart';
-
-class OwnershipController {
-  getOrAssignOnwership(DestinationModel destination, BuildContext context) {
-    TikiSdk tikiSdk =
-        Provider.of<WalletService>(context, listen: false).model.tikiSdk!;
-    Provider.of<OwnershipService>(context, listen: false)
-        .getOrAssignOwnership(destination.source, tikiSdk);
-  }
-}
