--- example/lib/ownership/model.dart
@@ -1,5 +1,0 @@
-import 'package:tiki_sdk_flutter/main.dart';
-
-class OwnershipExampleModel {
-  OwnershipModel? ownership;
-}
