+++ example/lib/request.dart
@@ -1,11 +1,11 @@
 import 'package:intl/intl.dart';
 
-class RequestModel {
+class Request {
   late final String timeStamp;
   final String icon;
   final String message;
 
-  RequestModel(this.icon, this.message) {
+  Request(this.icon, this.message) {
     timeStamp = DateFormat('hh:mm:ss').format(DateTime.now());
   }
 }
