--- example/lib/requests/model.dart
@@ -1,5 +1,0 @@
-import 'req_model.dart';
-
-class RequestsModel {
-  List<RequestModel> log = [];
-}
