--- example/lib/wallet/model.dart
@@ -1,9 +1,0 @@
-import 'package:tiki_sdk_flutter/main.dart';
-
-class WalletModel {
-  final String origin = "com.mytiki.tiki_sdk_example";
-  final String apiId = "2b8de004-cbe0-4bd5-bda6-b266d54f5c90";
-
-  TikiSdk? tikiSdk;
-  List<String> wallets = [];
-}
