+++ example/lib/widgets/body_card.dart
@@ -1,19 +1,12 @@
-import 'package:example/destination/service.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
-import '../requests/service.dart';
-import 'layout_edit.dart';
 
-class DestinationLayoutBtn extends StatelessWidget {
-  const DestinationLayoutBtn({super.key});
+class DestinationLayoutBodyBtn extends StatelessWidget {
+  const DestinationLayoutBodyBtn({super.key});
 
   @override
   Widget build(BuildContext context) {
-    DestinationService service =
-        Provider.of<DestinationService>(context, listen: true);
-    RequestsService requestsService =
-        Provider.of<RequestsService>(context, listen: false);
     return Container(
         padding: const EdgeInsets.symmetric(horizontal: 16.0, vertical: 8.0),
         decoration: BoxDecoration(
@@ -27,19 +20,20 @@ class DestinationLayoutBtn extends StatelessWidget {
                   children: const [
                     Expanded(
                         child: Text(
-                      "Destination",
+                      "Body",
                       style: TextStyle(fontWeight: FontWeight.bold),
                     )),
                     Icon(Icons.keyboard_arrow_right)
                   ]),
               const Padding(padding: EdgeInsets.all(8.0)),
-              Text("${service.model.httpMethod} ${service.model.url}")
+              Text(
+                service.model.body,
+                maxLines: 1,
+                overflow: TextOverflow.ellipsis,
+              )
             ]),
             onTap: () => Navigator.of(context).push(MaterialPageRoute(
-                builder: (context) => MultiProvider(providers: [
-                      ChangeNotifierProvider<RequestsService>.value(
-                          value: requestsService),
-                      ChangeNotifierProvider.value(value: service)
-                    ], child: DestinationLayoutEdit())))));
+                builder: (context) => ChangeNotifierProvider.value(
+                    value: service, child: DestinationLayoutBodyEdit())))));
   }
 }
