+++ example/lib/widgets/consent_detail.dart
