+++ example/lib/widgets/destination_card.dart
@@ -2,15 +2,18 @@ import 'package:example/destination/service.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
-import 'layout_body_edit.dart';
+import '../requests/service.dart';
+import '../destination/destination_edit.dart';
 
-class DestinationLayoutBodyBtn extends StatelessWidget {
-  const DestinationLayoutBodyBtn({super.key});
+class DestinationLayoutBtn extends StatelessWidget {
+  const DestinationLayoutBtn({super.key});
 
   @override
   Widget build(BuildContext context) {
     DestinationService service =
         Provider.of<DestinationService>(context, listen: true);
+    RequestsService requestsService =
+        Provider.of<RequestsService>(context, listen: false);
     return Container(
         padding: const EdgeInsets.symmetric(horizontal: 16.0, vertical: 8.0),
         decoration: BoxDecoration(
@@ -24,20 +27,19 @@ class DestinationLayoutBodyBtn extends StatelessWidget {
                   children: const [
                     Expanded(
                         child: Text(
-                      "Body",
+                      "Destination",
                       style: TextStyle(fontWeight: FontWeight.bold),
                     )),
                     Icon(Icons.keyboard_arrow_right)
                   ]),
               const Padding(padding: EdgeInsets.all(8.0)),
-              Text(
-                service.model.body,
-                maxLines: 1,
-                overflow: TextOverflow.ellipsis,
-              )
+              Text("${service.model.httpMethod} ${service.model.url}")
             ]),
             onTap: () => Navigator.of(context).push(MaterialPageRoute(
-                builder: (context) => ChangeNotifierProvider.value(
-                    value: service, child: DestinationLayoutBodyEdit())))));
+                builder: (context) => MultiProvider(providers: [
+                      ChangeNotifierProvider<RequestsService>.value(
+                          value: requestsService),
+                      ChangeNotifierProvider.value(value: service)
+                    ], child: DestinationLayoutEdit())))));
   }
 }
