+++ example/lib/widgets/destination_edit.dart
@@ -3,7 +3,7 @@ import 'package:example/requests/service.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
-import 'model.dart';
+import '../model.dart';
 
 class DestinationLayoutEdit extends StatelessWidget {
   DestinationLayoutEdit({super.key});
