+++ example/lib/widgets/ownership_detail.dart
